+++ pyramus/src/main/java/fi/otavanopisto/pyramus/views/students/ViewStudentViewController.java
@@ -67,7 +67,6 @@ import fi.otavanopisto.pyramus.domainmodel.users.User;
 import fi.otavanopisto.pyramus.domainmodel.users.UserVariable;
 import fi.otavanopisto.pyramus.framework.PyramusViewController2;
 import fi.otavanopisto.pyramus.framework.UserUtils;
-import fi.otavanopisto.pyramus.rest.controller.permissions.OrganizationPermissions;
 import fi.otavanopisto.pyramus.security.impl.PermissionController;
 import fi.otavanopisto.pyramus.util.StringAttributeComparator;
 import net.sf.json.JSONArray;
@@ -95,7 +94,7 @@ public class ViewStudentViewController extends PyramusViewController2 implements
     if (!PermissionController.instance().hasEnvironmentPermission(user, PyramusUIPermissions.VIEW_STUDENT)) {
       return false;
     } else {
-      if (PermissionController.instance().hasEnvironmentPermission(user, OrganizationPermissions.ACCESS_ALL_ORGANIZATIONS)) {
+      if (UserUtils.canAccessAllOrganizations(user)) {
         return true;
       } else {
         Long personId = requestContext.getLong("person");
@@ -165,7 +164,7 @@ public class ViewStudentViewController extends PyramusViewController2 implements
     StaffMember staffMember = staffMemberDAO.findByPerson(person);
     pageRequestContext.getRequest().setAttribute("staffMember", staffMember);
     
-    List<Student> students = PermissionController.instance().hasEnvironmentPermission(loggedUser, OrganizationPermissions.ACCESS_ALL_ORGANIZATIONS) ?
+    List<Student> students = UserUtils.canAccessAllOrganizations(loggedUser) ?
         studentDAO.listByPerson(person) : studentDAO.listByPersonAndOrganization(person, loggedUser.getOrganization());
     
     Collections.sort(students, new Comparator<Student>() {
